feat: add shared raster reader helpers
